moved files out of containers directory
