+++ src/layouts/index.js
@@ -5,9 +5,9 @@ import graphql from 'graphql'
 import Link, { withPrefix } from 'gatsby-link'
 import Helmet from 'react-helmet'
 import { ThemeProvider } from 'styled-components'
-import ContactSection from '../containers/Contact'
-import Footer from '../containers/Footer'
-import NavContainer from '../containers/Navbar'
+import ContactSection from '../components/Contact'
+import Footer from '../components/Footer'
+import NavContainer from '../components/Navbar'
 
 // code syntax-highlighting theme
 // feel free to change it to another one
